CDPlayProbe v7: track switching and natural end-of-track, the last two gaps
Both have been open since the beginning and neither could run, because the probe
always played the FIRST audio track from its START and always stopped after twelve
seconds. So a game changing tracks, and a game waiting for a track to end, were
untested paths in a project whose gate is "the music must start every time".

Phase B issues a second AudioPlay for a different track and polls six seconds. The
pump has to recompute gTrackStartLBA and the play range; that code has never
executed on hardware.

Phase C reaches a track boundary without sitting through a three-minute track: it
starts playback six seconds BEFORE the end, since the pump derives the end of the
range from the TOC and the boundary arrives on schedule either way. It then
watches for the status byte reaching 0x13 and the position ceasing to advance,
which is what a looping game polls for. That path has never run either.

Both phases skip loudly when the disc has no second audio track with a successor
in the TOC, and both reuse the position encoding phase A discovered rather than
running against a guess.

While writing phase C I put a "%ld" in a log line with no argument - in a
diagnostic whose whole job is to be trustworthy. Rather than just fixing it, the
three variadic logging calls now carry
__attribute__((format(printf,...))), so the compiler checks every format string in
the project. It immediately found a real latent defect in cd_recon.c, where a
volume's file count was passed as int to a %ld, and one zero-length format string.
Both fixed; all six build directories compile clean.

+++ probes/cdrecon/cd_recon.c
@@ -199,7 +199,7 @@ static void SurveyVolumes(void)
                index, cname, hpb.volumeParam.ioVRefNum,
                hpb.volumeParam.ioVDrvInfo, hpb.volumeParam.ioVDRefNum,
                (unsigned short)hpb.volumeParam.ioVSigWord,
-               hpb.volumeParam.ioVFSID, hpb.volumeParam.ioVNmFls);
+               hpb.volumeParam.ioVFSID, (long)hpb.volumeParam.ioVNmFls);
 
         if (hpb.volumeParam.ioVDRefNum != gR.cd.refNum) continue;
